requirements.txt
tensorflow_gpu==2.3.1
~umpy==1.21.5
tflite_support==0.3.1
opencv_python_headless==4.5.5.64
Flask==2.0.1
pandas==1.1.4
numpy==1.22.3
Pillow==9.1.0
secrets==1.0.2
tensorflow==2.8.0
